ints to longs Update seed parser to read longer numbers 1* solution

+++ Day 5 - Maps of Seeds/ConversionMap.cs
@@ -15,7 +15,7 @@ namespace Day5
             this.maps = maps;
             this.name = name;
         }
-        public int Map(int seed)
+        public long Map(long seed)
         {
             foreach(Map map in maps)
             {
